Fixed static analysis
