+++ analysis_options.yaml
@@ -24,8 +24,7 @@ linter:
     - use_key_in_widget_constructors: false
     - always_declare_return_types: true
     - await_only_futures: true
-    # avoid_print: false  # Uncomment to disable the `avoid_print` rule
-    # prefer_single_quotes: true  # Uncomment to enable the `prefer_single_quotes` rule
+    - avoid_print: true
 
 # Additional information about this file can be found at
 # https://dart.dev/guides/language/analysis-options
