+++ lib/Widgets/ImageHelper.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+
+class ImageHelper extends StatelessWidget
+{
+  final String url;
+  final String text;
+  ImageHelper(this.url, this.text);
+
+  Widget build(BuildContext context)
+  {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.center,
+      mainAxisAlignment: MainAxisAlignment.start,
+      children: [
+        Flexible(child: Container( height: 75, width:75, child: Image.network(url))),
+        Flexible(child: Text(text, textAlign: TextAlign.center,))
+      ],
+    );
+  }
+}
